+++ src/app/api/[[...route]]/kategori-asset.ts
@@ -4,6 +4,7 @@ import { zValidator } from '@hono/zod-validator';
 import { CreateKategoriAssetSchema, UpdateKategoriAssetSchema } from "@/schemas";
 import { z } from "zod";
 import { getKategoriAssetById, getKategoriAssetBykode } from "@/data/kategori-asset";
+import { auth } from "@/auth";
 
 
 const app = new Hono()
@@ -44,6 +45,12 @@ const app = new Hono()
         zValidator("json", CreateKategoriAssetSchema),
         async (c) => {
 
+            const session = await auth();
+
+            if (!session) {
+                return c.json({ error: "Unauthorized" }, 401);
+            }
+
             const values = c.req.valid("json");
             const { kode } = values;
             const kategoriAsset = await getKategoriAssetBykode(kode);
@@ -71,6 +78,12 @@ const app = new Hono()
             }),
         ),
         async (c) => {
+            const session = await auth();
+
+            if (!session) {
+                return c.json({ error: "Unauthorized" }, 401);
+            }
+
             const values = c.req.valid("json");
 
             const data = await db.kategoriAsset.deleteMany({
@@ -89,6 +102,12 @@ const app = new Hono()
         })),
         zValidator("json", UpdateKategoriAssetSchema),
         async (c) => {
+            const session = await auth();
+
+            if (!session) {
+                return c.json({ error: "Unauthorized" }, 401);
+            }
+
             const { id } = c.req.valid("param");
             const values = c.req.valid("json");
 
@@ -112,6 +131,12 @@ const app = new Hono()
             id: z.string().optional(),
         })),
         async (c) => {
+            const session = await auth();
+
+            if (!session) {
+                return c.json({ error: "Unauthorized" }, 401);
+            }
+
             const { id } = c.req.valid("param");
 
             if (!id) {
